requirements.txt
requests
gevent
Flask
loguru
Werkzeug>=0.16.1
tmdbsimple
PyYAML
pycryptodome==3.12.0
arrapi